+++ Sources/DDDCore/Entity/AggregateRootMetadata.swift
@@ -11,7 +11,7 @@ public class AggregateRootMetadata {
     var events: [any DomainEvent] = []
 
     public package(set) var deleted: Bool
-    public package(set) var version: UInt?
+    public package(set) var version: UInt64?
 
     public init() {
         deleted = false
